+++ lib/Params.ml
@@ -308,13 +308,26 @@ module Exp = struct
   let box_fun_decl ~ctx0 c k =
     match ctx0 with
     | _ when ocp c -> hvbox 2 k
+    (* Avoid large indentation for [let _ = function]. *)
+    | Lb
+        {pvb_body= Pfunction_body {pexp_desc= Pexp_function ([], _, _); _}; _}
+      ->
+        hovbox 2 k
     | Str _ | Lb _ | Clf _ | Exp {pexp_desc= Pexp_let _; _} -> hovbox 4 k
     | _ -> hvbox 2 k
 
-  let box_pro_with_match ~ctx0 ~parens =
+  let match_inner_pro ~ctx0 ~parens =
     if parens then false
     else
       match ctx0 with Exp {pexp_desc= Pexp_infix _; _} -> false | _ -> true
+
+  let function_inner_pro ~has_cmts_outer ~ctx0 =
+    if has_cmts_outer then false
+    else
+      match ctx0 with
+      | Str _ | Lb _ | Exp {pexp_desc= Pexp_ifthenelse _ | Pexp_let _; _} ->
+          false
+      | _ -> true
 end
 
 module Mod = struct