+++ CHANGES.md
@@ -33,8 +33,8 @@ profile. This started with version 0.26.0.
 
 ### Changed
 
-- `begin match` can now be printed on the same line, with one less indentation
-  level for the body of the match. (#2666, @EmileTrotignon)
+- `begin match` and `begin fun` can now be printed on the same line, with one less indentation
+  level for the body of the inner expression. (#2666, @EmileTrotignon)
 
 ## 0.27.0